honeybee: Moved font and graphic modules to ntcip sub-module.
