+++ bin/honeybee/src/ntcip/mod.rs
@@ -2,14 +2,7 @@
 //
 // Copyright (C) 2019  Minnesota Department of Transportation
 //
-// This program is free software; you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation; either version 2 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
+mod base64;
+pub mod font;
+pub mod graphic;
 pub mod multi;
